Moved checkstyle to its own package
