+++ test/handlers/events/delivery/review/checkstyle/checkstyleReportToReviewTest.ts
@@ -3,8 +3,8 @@ import "mocha";
 import * as assert from "power-assert";
 
 import { SimpleRepoId } from "@atomist/automation-client/operations/common/RepoId";
-import { extract } from "../../../../../src/handlers/events/delivery/review/checkstyleReportExtractor";
-import { checkstyleReportToReview } from "../../../../../src/handlers/events/delivery/review/checkStyleReportToReview";
+import { extract } from "../../../../../../src/handlers/events/delivery/review/checkstyle/checkstyleReportExtractor";
+import { checkstyleReportToReview } from "../../../../../../src/handlers/events/delivery/review/checkstyle/checkStyleReportToReview";
 import { xml2valid1 } from "./checkstyleReportExtractorTest";
 
 describe("checkstyleReportToReview", () => {
