+++ test/handlers/events/delivery/review/checkstyle/checkstyleReviewerTest.ts
@@ -3,7 +3,7 @@ import "mocha";
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
 import { NodeFsLocalProject } from "@atomist/automation-client/project/local/NodeFsLocalProject";
 import * as assert from "power-assert";
-import { checkstyleReviewer } from "../../../../../src/handlers/events/delivery/review/checkstyleReviewer";
+import { checkstyleReviewer } from "../../../../../../src/handlers/events/delivery/review/checkstyle/checkstyleReviewer";
 
 const checkstylePath = process.env.CHECKSTYLE_PATH;
